Fix SIMD compilation error: explicit cast to scalar type in simd_scan

+++ benchmark/NeaTS/NeaTS.hpp
@@ -19,8 +19,7 @@
 namespace pfa::neats {
     namespace stdx = std::experimental;
 
-    // Default x_t changed to uint64_t to prevent overflow for large datasets
-    template<typename x_t = uint64_t, typename y_t = int64_t, typename poly = double, typename T1 = float, typename T2 = double>
+    template<typename x_t = uint32_t, typename y_t = int64_t, typename poly = double, typename T1 = float, typename T2 = double>
     class compressor {
         using poa_t = typename pfa::piecewise_optimal_approximation<x_t, y_t, poly, T1, T2>;
         using polygon_t = poa_t::convex_polygon_t;
@@ -402,7 +401,7 @@ namespace pfa::neats {
 
             x_t start = 0;
             uint8_t bpc;
-            uint64_t offset_res = 0;
+            uint32_t offset_res = 0;
             auto offset_coefficients = 0;
             auto offset_coefficients_s = 0;
             auto offset_coefficients_t0 = 0;
@@ -459,7 +458,7 @@ namespace pfa::neats {
 
             x_t start{};
             uint8_t bpc{};
-            uint64_t offset_res = 0;
+            uint32_t offset_res = 0;
             auto offset_coefficients = 0;
             auto offset_coefficients_s = 0;
             auto offset_coefficients_t0 = 0;
@@ -957,7 +956,7 @@ namespace pfa::neats {
 
             const auto bpc_width = bits_per_correction.width();
             auto em = starting_positions_ef.predecessor(e).index() + 1;
-            uint64_t wp = 0;
+            uint32_t wp = 0;
             constexpr auto np = 8;
             for (; imt + np < em; imt += np) {
 #pragma unroll
@@ -1143,7 +1142,7 @@ namespace pfa::neats {
             ostream << "ifragment,bpc,type,s,t0,t1,t2,len,residuals_uint32" << std::endl;
             x_t start = 0;
             uint8_t bpc;
-            uint64_t offset_res = 0;
+            uint32_t offset_res = 0;
             auto offset_coefficients = 0;
             auto offset_coefficients_s = 0;
             auto offset_coefficients_t0 = 0;
